Use chrpath to fix incorrectly escaped rpath
We escape rpath incorrectly when setting it in LDFLAGS, resulting
in it being set to something like 'RIGIN/../lib', when it should
be '$ORIGIN/../lib'

This patch solves the problem by setting it to XORIGIN in LDFLAGS
instead of $ORIGIN and then updating it to the correct value with
chrpath in finalize_package_build. XORIGIN is used because chrpath
can't change the length of the rpath and its something that won't
have any escaping problems.

Change-Id: I6ee0e67febac1df5b867e297a506a6511bf2463c